requirements.txt
pandas==2.0.3
sqlalchemy==2.0.20
numpy==1.24.4
pytest==7.4.0
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
psycopg2-binary==2.9.7  
